requirements.txt
apache-airflow-providers-google
apache-airflow-providers-http
pyarrow
apache-airflow[google]